requirements.txt
asgiref==3.7.2
click==8.1.7
coverage==7.3.2
Django==4.2.7
django-extensions==3.2.3
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
iniconfig==2.0.0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==7.4.3
pytz==2023.3.post1
sqlparse==0.4.4
svgwrite==1.4.3
